Struct interner::global::GlobalPool

pub struct GlobalPool<T, S = RandomState>(/* private fields */)
where
    T: Poolable + Debug + Clone + Eq + PartialEq + Hash + Ord + PartialOrd + 'static,
    S: BuildHasher + 'static;

A global interned pool.

This type is used to create globally allocated pools.

use interner::global::{GlobalPool, GlobalString};

static GLOBAL_STRINGS: GlobalPool<String> = GlobalPool::new();

let interned = GLOBAL_STRINGS.get(String::from("hello"));
let second = GLOBAL_STRINGS.get("hello");

assert!(GlobalString::ptr_eq(&interned, &second));

Implementations

impl<T> GlobalPool<T>
where T: Poolable + Debug + Clone + Eq + PartialEq + Hash + Ord + PartialOrd,

pub const fn new() -> Self

Returns a new instance using RandomState for the internal hashing.

impl<T, S> GlobalPool<T, S>
where T: Poolable + Debug + Clone + Eq + PartialEq + Hash + Ord + PartialOrd, S: BuildHasher,

pub fn pooled<C>(&'static self) -> C
where C: FromIterator<Pooled<&'static Self, S>>,

Returns a collection of the currently pooled items.

impl<T, S> GlobalPool<T, S>
where T: Poolable + Debug + Clone + Eq + PartialEq + Hash + Ord + PartialOrd, S: BuildHasher,

pub const fn with_hasher(hasher: S) -> Self

Returns a new instance using the provided hasher.

pub const fn with_hasher_init(init: fn() -> S) -> Self

Returns a new instance using the function to load the hasher when the pool is initialized on first use.

pub const fn with_capacity_and_hasher(capacity: usize, hasher: S) -> Self

Returns a new instance using the provided hasher with enough capacity to hold the requested number of items without reallocating.

pub const fn with_capacity_and_hasher_init( capacity: usize, init: fn() -> S, ) -> Self

Returns a new instance using the function to load the hasher when the pool is initialized on first use. The returned instance has enough capacity to hold the requested number of items without reallocating.

impl<S> GlobalPool<String, S>
where S: BuildHasher,

pub fn get<'a, V>(&'static self, value: V) -> GlobalString<S>
where V: Into<Cow<'a, str>>,

Returns a copy of an existing GlobalString if one is found. Otherwise, a new GlobalString is created and returned.

While any copies of the returned GlobalString are still allocated, calling this function is guaranteed to return a copy of the same string.

pub const fn get_static( &'static self, value: &'static str, ) -> StaticPooledString<S>

Returns a static pooled string, which keeps the pooled string allocated for the duration of the process.

The string is not initialized until it is retrieved for the first time.

pub const fn get_static_with( &'static self, function: fn() -> Cow<'static, str>, ) -> StaticPooledString<S>

Returns a static pooled string, which keeps the pooled string allocated for the duration of the process. The string is initialized using the function provided when it is retrieved for the first time.

impl<S> GlobalPool<PathBuf, S>
where S: BuildHasher,

pub fn get<'a, V>(&'static self, value: V) -> GlobalPath<S>
where V: Into<Cow<'a, Path>>,

Returns a copy of an existing GlobalPath if one is found. Otherwise, a new GlobalPath is created and returned.

While any copies of the returned GlobalPath are still allocated, calling this function is guaranteed to return a copy of the same path.

pub const fn get_static_with( &'static self, function: fn() -> Cow<'static, Path>, ) -> StaticPooledPath<S>

Returns a static pooled path, which keeps the pooled path allocated for the duration of the process. The path is initialized using the function provided when it is retrieved for the first time.

impl<S> GlobalPool<Vec<u8>, S>
where S: BuildHasher,

pub fn get<'a, V>(&'static self, value: V) -> GlobalBuffer<S>
where V: Into<Cow<'a, [u8]>>,

Returns a copy of an existing GlobalBuffer if one is found. Otherwise, a new GlobalBuffer is created and returned.

While any copies of the returned GlobalBuffer are still allocated, calling this function is guaranteed to return a copy of the same byte buffer.

pub const fn get_static( &'static self, value: &'static [u8], ) -> StaticPooledBuffer<S>

Returns a static pooled buffer, which keeps the pooled buffer allocated for the duration of the process.

The buffer is not initialized until it is retrieved for the first time.

pub const fn get_static_with( &'static self, function: fn() -> Cow<'static, [u8]>, ) -> StaticPooledBuffer<S>

Returns a static pooled buffer, which keeps the pooled buffer allocated for the duration of the process. The buffer is initialized using the function provided when it is retrieved for the first time.